In attempting to address a particularly nasty Enum parsing issue, I fixed several minor issues in ScalaBuff:
  1. Fix the 'update-test-resources' code so that the generated files are
     deleted BEFORE they are rewritten, not AFTER.

  2. Update to the latest ScalaTest version, avoiding deprecation warnings.

  3. Fixed broken tests with a clean checkout.

And after getting those polished, I was able to write regression tests for
the actual bug to fix.  Notably, when a message is parsing an OPTIONAL enum
field, if an UnknownEnumException is thrown, AND there is a default value
for that option, then fall back to using that default value rather.  If
there is no default, then continue throwing that exception.